@@ -1,26 +1,17 @@
 {
-  "name": "eesayas.com",
+  "name": "isaiasbriones.com",
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "serve": "vue-cli-service serve",
-    "build": "vue-cli-service build",
-    "start": "vue-cli-service build && serve -s dist -l 3003",
-    "lint": "vue-cli-service lint"
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview"
   },
   "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "^1.2.32",
-    "@fortawesome/free-brands-svg-icons": "^5.15.1",
-    "@fortawesome/free-solid-svg-icons": "^5.15.1",
-    "@fortawesome/vue-fontawesome": "^2.0.0",
-    "bootstrap": "^4.5.3",
-    "bootstrap-vue": "^2.20.1",
-    "core-js": "^3.6.5",
-    "vue": "^2.6.11"
-  },
-  "devDependencies": {
-    "@vue/cli-service": "^3.12.1",
-    "vue-template-compiler": "^2.6.11"
+    "@vitejs/plugin-vue": "^5.2.1",
+    "bootstrap": "^5.3.3",
+    "vite": "^6.0.7",
+    "vue": "^3.5.13"
   },
   "browserslist": [
     "> 1%",
