@@ -1,32 +1,8 @@
-import Vue from 'vue'
-import App from './App.vue'
-import { BootstrapVue, IconsPlugin } from 'bootstrap-vue'
-import { library } from '@fortawesome/fontawesome-svg-core'
-import { faGithub, faLinkedin } from '@fortawesome/free-brands-svg-icons'
-import { faFile, faCode, faDesktop, faEnvelope, faPhone } from '@fortawesome/free-solid-svg-icons'
-import { FontAwesomeIcon } from '@fortawesome/vue-fontawesome'
+import { createApp } from "vue";
+import App from "./App.vue";
+import "./style.css";
 
-library.add(faGithub)
-library.add(faLinkedin)
-library.add(faFile)
-library.add(faCode)
-library.add(faDesktop)
-library.add(faEnvelope)
-library.add(faPhone)
+import "bootstrap/dist/css/bootstrap.css";
 
-Vue.component('font-awesome-icon', FontAwesomeIcon)
-
-// Install BootstrapVue
-Vue.use(BootstrapVue)
-
-// Optionally install the BootstrapVue icon components plugin
-Vue.use(IconsPlugin)
-
-import 'bootstrap/dist/css/bootstrap.css'
-import 'bootstrap-vue/dist/bootstrap-vue.css'
-
-Vue.config.productionTip = false
-
-new Vue({
-  render: h => h(App),
-}).$mount('#app')
+const app = createApp(App);
+app.mount("#app");
